Add: 그룹 생성페이지 색깔 클릭하여 기능구현 - 조성윤 2025-02-12 23:04 Refactor: 식당 상세페이지 리팩토링 - 조성윤 2025-02-12 23:04

+++ Prepay/app/src/main/java/com/example/prepay/ui/RestaurantDetails/OrderHistoryViewModel.kt
@@ -15,7 +15,7 @@ class OrderHistoryViewModel : ViewModel() {
     val orderHistoryListInfo : LiveData<List<OrderHistory>>
         get() = _orderHistoryListInfo
 
-    fun getAllOrderHistoryList(userId:Int, teamId:Long, storeId : Long) {
+    fun getAllOrderHistoryList(userId:Int, teamId:Long, storeId: Int) {
         viewModelScope.launch {
             runCatching {
                 RetrofitUtil.orderService.getDetailHistory(userId, OrderHistoryReq(teamId, storeId))
